Refactor analysis module structure
- Renamed Analysis.py to reflect its new location in the job.analysis directory.
- Updated import paths in job and test files to match the new structure.
