+++ src/easydiffraction/job/analysis/analysis.py
